// src/types/traffics.ts
export interface Traffic {
    no: number;
    ruas: string;
    gerbang: string;
    gardu: number;
    tanggal: string;
    hari: string;
    payment_method: string;
    gol_1: number;
    gol_2: number;
    gol_3: number;
    gol_4: number;
    gol_5: number;
    total: number;
    id: number;
  }